updating code

+++ rocky-linux-hypervisor.sh
@@ -243,10 +243,13 @@ function configure_ansible_vault_setup() {
 
     
     source ~/.profile
-    curl -OL https://github.com/tosin2013/ansiblesafe/releases/download/v${ANSIBLE_SAFE_VERSION}/ansiblesafe-v${ANSIBLE_SAFE_VERSION}-linux-amd64.tar.gz
-    tar -zxvf ansiblesafe-v${ANSIBLE_SAFE_VERSION}-linux-amd64.tar.gz
-    chmod +x ansiblesafe-linux-amd64
-    sudo mv ansiblesafe-linux-amd64 /usr/local/bin/ansiblesafe
+    if [ ! -f /usr/local/bin/ansiblesafe ];
+    then
+        curl -OL https://github.com/tosin2013/ansiblesafe/releases/download/v${ANSIBLE_SAFE_VERSION}/ansiblesafe-v${ANSIBLE_SAFE_VERSION}-linux-amd64.tar.gz
+        tar -zxvf ansiblesafe-v${ANSIBLE_SAFE_VERSION}-linux-amd64.tar.gz
+        chmod +x ansiblesafe-linux-amd64
+        sudo mv ansiblesafe-linux-amd64 /usr/local/bin/ansiblesafe
+    fi 
 
     ansiblesafe -f /root/qubinode_navigator/inventories/localhost/group_vars/control/vault.yml
     generate_inventory /root
